Prevent LogsDB Index Template delete error.

+++ packages/kbn-apm-synthtrace/src/scenarios/logs_and_metrics.ts
@@ -39,7 +39,7 @@ const scenario: Scenario<LogDocument> = async (runOptions) => {
       if (isLogsDb) await logsEsClient.createIndexTemplate(IndexTemplateName.LogsDb);
     },
     teardown: async ({ logsEsClient }) => {
-      await logsEsClient.deleteIndexTemplate(IndexTemplateName.LogsDb);
+      if (isLogsDb) await logsEsClient.deleteIndexTemplate(IndexTemplateName.LogsDb);
     },
     generate: ({ range, clients: { logsEsClient, apmEsClient } }) => {
       const { numServices = 3 } = runOptions.scenarioOpts || {};
