+++ packages/kbn-apm-synthtrace/src/scenarios/logs_traces_hosts.ts
@@ -53,7 +53,7 @@ const scenario: Scenario<LogDocument | InfraDocument | ApmFields> = async (runOp
       if (isLogsDb) await logsEsClient.createIndexTemplate(IndexTemplateName.LogsDb);
     },
     teardown: async ({ logsEsClient }) => {
-      await logsEsClient.deleteIndexTemplate(IndexTemplateName.LogsDb);
+      if (isLogsDb) await logsEsClient.deleteIndexTemplate(IndexTemplateName.LogsDb);
     },
     generate: ({ range, clients: { logsEsClient, infraEsClient, apmEsClient } }) => {
       const {
